✨ search struct
