+++ frontend/src/lib/api.ts
@@ -1,17 +1,7 @@
 import axios from "axios";
+import { SearchResult } from "@/lib/structs.ts";
 
 const apiUrl = import.meta.env.VITE_API_URL;
-
-export interface SearchResult {
-  id: string;
-  entity_type: string;
-  provider_id: number;
-  provider_name: string;
-  recreation_area_id: string | null;
-  recreation_area_name: string | null;
-  campground_id: string | null;
-  campground_name: string | null;
-}
 
 // Configure axios with base URL
 const api = axios.create({
