+++ frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.test.tsx","./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/camplylogo.tsx","./src/components/dismissiblebanner.tsx","./src/components/header.tsx","./src/components/layout.tsx","./src/components/searchbar.tsx","./src/components/searchdropdown.tsx","./src/components/theme-provider.tsx","./src/components/theme-toggle.tsx","./src/components/ui/button-variants.ts","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/hooks/use-theme.ts","./src/hooks/usesearch.ts","./src/lib/api.ts","./src/lib/utils.ts","./src/pages/auth.tsx","./src/pages/campground.tsx","./src/pages/contact.tsx","./src/pages/contribute.tsx","./src/pages/ethos.tsx","./src/pages/faq.tsx","./src/pages/home.tsx","./src/pages/howitworks.tsx","./src/pages/privacypolicy.tsx","./src/pages/providers.tsx","./src/pages/recreationarea.tsx","./src/pages/termsofservice.tsx","./src/test/setup.ts"],"version":"5.6.3"}
+{"root":["./src/app.test.tsx","./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/camplylogo.tsx","./src/components/dismissiblebanner.tsx","./src/components/header.tsx","./src/components/layout.tsx","./src/components/map.tsx","./src/components/searchbar.tsx","./src/components/searchdropdown.tsx","./src/components/theme-provider.tsx","./src/components/theme-toggle.tsx","./src/components/ui/button-variants.ts","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/hooks/use-theme.ts","./src/hooks/usesearch.ts","./src/lib/api.ts","./src/lib/utils.ts","./src/pages/auth.tsx","./src/pages/campground.tsx","./src/pages/contact.tsx","./src/pages/contribute.tsx","./src/pages/ethos.tsx","./src/pages/faq.tsx","./src/pages/home.tsx","./src/pages/howitworks.tsx","./src/pages/privacypolicy.tsx","./src/pages/providers.tsx","./src/pages/recreationarea.tsx","./src/pages/termsofservice.tsx","./src/test/setup.ts"],"version":"5.6.3"}
